requirements.txt
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
colorama==0.4.6
h11==0.14.0
idna==3.6
keyboard==0.13.5
numpy==1.26.4
outcome==1.3.0.post0
pandas==2.2.1
pillow==10.2.0
playsound==1.3.0
pycparser==2.21
pygame==2.5.2
pypng==0.20220715.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
qrcode==7.4.2
selenium==4.18.1
setuptools==69.1.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wheel==0.42.0
wsproto==1.2.0
